# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Registro Elettronico V2 API
  description: API REST per la gestione scolastica (Auth, Utenti, Voti, Presenze, Comunicazioni, Documenti, Colloqui, Agenda).
  version: 0.8.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local Development Server
paths:
...
  /timetable/generate:
    post:
      summary: Avvia generazione orario scolastico automatica con risolutore CSP e priorità anzianità
      tags:
        - Orario Scolastico
      security:
        - BearerAuth: []
      responses:
        '202':
          description: Generazione orario avviata in background
  /timetable/generate/{jobID}:
    get:
      summary: Polling stato avanzamento calcolo orario
      tags:
        - Orario Scolastico
      security:
        - BearerAuth: []
      parameters:
        - name: jobID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Stato del job e orario proposto
  /timetable/generate/{jobID}/publish:
    post:
      summary: Pubblicazione orario calcolato nelle tabelle ufficiali
      tags:
        - Orario Scolastico
      security:
        - BearerAuth: []
      parameters:
        - name: jobID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Orario pubblicato con successo
  /timetable/preferences:
    get:
      summary: Recupera desiderata orario docente
      tags:
        - Orario Scolastico
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Desiderata docente
    post:
      summary: Salva desiderata orario docente
      tags:
        - Orario Scolastico
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Desiderata salvati con successo
  /timetable/constraints:
    get:
      summary: Lista vincoli orario istituto e plessi
      tags:
        - Orario Scolastico
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Elenco vincoli orari
    post:
      summary: Salva vincoli orario istituto
      tags:
        - Orario Scolastico
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Vincoli salvati con successo
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT